Security fixes v1.0.2 — address all audit findings
- Fix fatal error: find_chapter() → get_chapter() in emails
- Add sanitize_callback with hex validation to register_setting()
- Add capability check (edit_post) to woo save_product_field()
- Validate primary_color as hex before CSS injection in PDF
- Sanitize from_name/email in wp_mail() headers (prevent injection)
- Cast $pct to int in shortcode progress widget
- Add user existence check in admin_toggle_coach REST endpoint
- Replace echo <script> with wp_print_inline_script_tag()
- Escape CSS width calc with esc_attr/number_format in PDF

+++ includes/class-klarhed-shortcodes.php
@@ -33,7 +33,7 @@ class Klarhed_Shortcodes {
 
     public function progress() {
         $this->enqueue();
-        $pct = Klarhed_Progress::completion_pct();
+        $pct = (int) Klarhed_Progress::completion_pct();
         return '<div class="klarhed-progress-widget"><span>' . esc_html__( 'Dit forløb', 'klarhed' ) . '</span><div class="kh-progress"><div style="width:' . $pct . '%"></div></div><b>' . $pct . '%</b></div>';
     }
